restructure directory, add fsm transition test
