+++ src/software/ai/hl/stp/play/BUILD
@@ -208,28 +208,6 @@ cc_library(
 )
 
 cc_library(
-    name = "penalty_kick_play",
-    srcs = [
-        "penalty_kick_play.cpp",
-        "penalty_kick_play_fsm.cpp",
-    ],
-    hdrs = [
-        "penalty_kick_play.h",
-        "penalty_kick_play_fsm.h",
-    ],
-    deps = [
-        ":play",
-        "//shared:constants",
-        "//software/ai/hl/stp/tactic/goalie:goalie_tactic",
-        "//software/ai/hl/stp/tactic/move:move_tactic",
-        "//software/ai/hl/stp/tactic/penalty_kick:penalty_kick_tactic",
-        "//software/logger",
-        "//software/util/generic_factory",
-    ],
-    alwayslink = True,
-)
-
-cc_library(
     name = "shoot_or_chip_play",
     srcs = ["shoot_or_chip_play.cpp"],
     hdrs = ["shoot_or_chip_play.h"],
@@ -304,7 +282,6 @@ cc_library(
         ":kickoff_enemy_play",
         ":kickoff_friendly_play",
         ":penalty_kick_enemy_play",
-        ":penalty_kick_play",
         ":shoot_or_chip_play",
         ":stop_play",
         "//software/ai/hl/stp/play/crease_defense:crease_defense_play",
@@ -313,6 +290,7 @@ cc_library(
         "//software/ai/hl/stp/play/hardware_challenge_plays:scoring_from_contested_possession_play",
         "//software/ai/hl/stp/play/hardware_challenge_plays:scoring_with_static_defenders_play",
         "//software/ai/hl/stp/play/offense:offense_play",
+        "//software/ai/hl/stp/play/penalty_kick:penalty_kick_play",
         "//software/ai/hl/stp/play/shoot_or_pass:shoot_or_pass_play",
     ],
 )
@@ -478,22 +456,6 @@ cc_test(
         "//shared/test_util:tbots_gtest_main",
         "//software/ai/hl/stp/play:shoot_or_chip_play",
         "//software/simulated_tests:simulated_er_force_sim_play_test_fixture",
-        "//software/simulated_tests/validation:validation_function",
-        "//software/test_util",
-        "//software/time:duration",
-        "//software/world",
-    ],
-)
-
-cc_test(
-    name = "penalty_kick_play_test",
-    srcs = ["penalty_kick_play_test.cpp"],
-    deps = [
-        "//shared/test_util:tbots_gtest_main",
-        "//software/ai/hl/stp/play:penalty_kick_play",
-        "//software/simulated_tests:simulated_er_force_sim_play_test_fixture",
-        "//software/simulated_tests/non_terminating_validation_functions",
-        "//software/simulated_tests/terminating_validation_functions",
         "//software/simulated_tests/validation:validation_function",
         "//software/test_util",
         "//software/time:duration",
