+++ src/software/ai/hl/stp/play/penalty_kick/penalty_kick_play.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "proto/parameters.pb.h"
-#include "software/ai/hl/stp/play/penalty_kick_play_fsm.h"
+#include "software/ai/hl/stp/play/penalty_kick/penalty_kick_play_fsm.h"
 #include "software/ai/hl/stp/play/play.h"
 
 /**
