+++ src/software/ai/play_selection_fsm.cpp
@@ -9,8 +9,8 @@
 #include "software/ai/hl/stp/play/kickoff_enemy_play.h"
 #include "software/ai/hl/stp/play/kickoff_friendly_play.h"
 #include "software/ai/hl/stp/play/offense/offense_play.h"
+#include "software/ai/hl/stp/play/penalty_kick/penalty_kick_play.h"
 #include "software/ai/hl/stp/play/penalty_kick_enemy_play.h"
-#include "software/ai/hl/stp/play/penalty_kick_play.h"
 #include "software/ai/hl/stp/play/stop_play.h"
